Converted stream functions in job stores to context managers
This is necessary because the AWS job store needs to get control back when the file handle is closed. By turning the ...Stream() methods into CMs, we can ensure that the stream is closed *and* do the necessary book-keeping afterwards. In the case of the AWS jobs store, the book-keeping is mainly the registration of the resulting S3 file version ID in the SimpleDB database.