Do not delete job store on exception during creation
Unfortunately, jobTree passes a new config, even on resumes. Since we derive the create flag from the presence of the config parameter, create is True during resumes. We don't want to a resume to delete the job store, even if something goes wrong. What motivated this change was the fact that I pressed Ctrl-C during a test run. The Ctrl-C happened to occur in the try block of the constructor during a resume and caused the job store to be deleted. This can be reintroduced when we have a more reliable way to detect the initial invocation.