Streamlined setup.py and Makefile (resolves #207)
This commit eliminates the need for setting PYTHONPATH in the Makefile or a developer's .profile by replacing it with setuptools' develop mode. But instead of using "setup.py develop" directly, we use "pip install -e" as a frontend. This allows for the use of setuptools' extras feature (see below) in conjunction with develop mode. Pip calls that "editable install". To perform the editable install, simply run "make develop" or just "make". To run tests, use "make test" as usual. Running "make clean" undoes develop mode.

By default, "make develop" installs into the user-specific install location (via "pip install --user") so as to not require root privileges. However, it is smart enough to omit "--user "when it detects that it is running in a virtualenv. This by the way is the recommended way to develop with Toil: create a virtualenv, activate it and then run "make".

Setup.py now specifies "extras". With this mechanism, a user can now chose whether to install Toil with AWS and/or Mesos support: "pip install toil[aws,mesos]" or "pip install toil".