Fixed pickling of service jobs by importing and hot-deploying corresponding user module
Cherry picked 0023582d28cdfea82940734a8d1f50005a46bbe8 (+3 squashed commits)
Squashed commits:
[c89fac4] Fixes to hot deployment of service- and function-wrapping jobs

Only localize user function modules if they don't belong to toil. Ensure that the module containing a service whose job is at the root of a workflow will be hot-deployed.
[98a833c] Added missing globalization of service module descriptor
[e9e990a] Utils test was broken but the error was masked because sortTest success
was set to equal 1. This is fixed, as is the serviceTestIssue. Hannes,
please test this now works for you on pyCharm.