Unicode-encode log messages
ElementTree tries to encode the log messages with XML entities, and it
can be unable to do this if bytes >128 are present in `str` values in
the tree.

I have solved this by converting all the relevant strings to `unicode`
objects, assuming that the user script output being collected is
actually UTF-8. In the event of a UTF-8 decoding error, the offending
characters are replaced with placeholders, but the Toil code should keep
going (and ElementTree should still be able to serialize the logs).