Fix slow restart against the AzureJobStore
Basically, we now download all the jobs from the jobStore before trying
to "clean" the jobStore or rebuild the ToilState. This means we aren't
running individual per-job operations against every job in the jobStore,
which can be slow if the jobStore is Azure or AWS and each of those
operations requires a round-trip HTTP request.

To make this work, I also fixed the Azure job store's `jobs()` method,
which was only returning the first 1000 jobs.

The `clean()` method and the restart process in general is now much
noisier, with info messages letting you know how many jobs have been
loaded so far.

Fix logger naming in AzureJobStore