Allow uploading unowned files to file store
This commit checks file ownership before trying to change the
permissions on a file that a job wants to put in the file store. If you
want to, say, store the local node's syslog file, or a file owned by
another user on a shared disk, we have to do this.