Changes wrt backwards compatibility (--backwardsCompatible=True)
1. Added a --backwardsCompatible input flag that allows old scripts to
run with the new caching schema. This flag is used when readGlobalFile
is callled.
2. IMPORTANT: --backwardsCompatible should default to False but for this
commit it is True to show that none of the FileStore functions have been
messed up.
3. src/toil/jobStores/fileJobStore.py will link to the job store file
even in the non-shared caching system if --backwardsCompatible was set
to
False.
4. For all future versions of Toil, files obtained by readGlobalFile are
going to be immutable, unless the user specifically requests them with
the modifiable argument.
5. Modified sort.py and to not modify the files downloaded with
readGlobalFile directly. Also, sort.py:cleanup now moves the file to the
desired location explicitly with shutil.copyfile since job cleanup will
delete all files read with readGlobalFile.
6. Fixed teh bug that left behind a file during the last jenkins build.
7. modified .gitignore to ignore the .idea directory and it's contents.