Major changes to caching
1. Completely reworked how the cache is set up. Previously, if a worker
dies while trying to set up the cache for the first time, all subsequent
workers would go into an infinite cycle of timing out waiting for the
completion of this killed job. The new system  uses a setup lock to
ensure graceful setup of the cache.
2. Resource.py now has a cleanCache function (possibly needs a better
name so as not to cloash with CachedFileStore.cleanCache) that will
purge the cache directory when called.
3. the SingleMachine batch system now calls resource.Resource.cleanCache
on shutdown. Need to figure out how this will work with other batch
systems.
4. jobCacheTest.py modified to be compatible with (2) and (3).
5. Modified how jobStoreTest checks the NoSuchFileException so it
consistently gets caught across job store types. @hannes-ucsc can you
see if this is ok?