Add logging when rejecting a resource offer (resolves #669)
Currently, when we reject a resource offer because it does not satisfy the
resource requirements for any of the currently offered tasks, we only
log at debug level. Additionally, we don't explain why the offer was rejected,
and we only explain for the job with the fewest cores requested.

This commit checks to see if we are rejecting a resource offer while we do not
have any jobs currently running (i.e., we are hung/not making progress), and
logs the reason that the currently offered jobs are not being scheduled at the
WARNING logging level.