Use the env variable TOIL_WORKDIR if workdir is not provided (resolves #691)
resolves #691

Created a new function in common.py that returns the path to the workflow-specific
directory based on the value of workDir passed to Toil. The directory is one level
below workDir and is named toil-<workflowID>.

If workDir was not passed, look for TOIL_WORKDIR in the environment.
If TOIL_WORKDIR is not in the environment, then use the machine's
temp directory (obtained from tempfile module).

This directory is cleaned up in the worker(supported by single machine and mesos).