Use ephemeral for running Toil tests on Jenkins (resolves #704)
resolves #704

The test framework uses the system TMPDIR to create temp working directories for running toil tests. If we
set it to a directory on /mnt/ephemeral while running tests on jenkins, then the tests can run in a larger volume
allowing for tests to use more space if need be.