Add support for launching Spark Clusters as a service (resolves #774)
Adds a package `toil.spark` that contains a module that launches Spark clusters
using the `toil.job.ServiceJob` infrastructure. This is a refactoring of the
code that currently exists in toil-scripts. These clusters are resilient to
node failures provided that they are launched from a checkpointable job.

H/T @almussel, who contributed heavily to this work.