Add YARNBatchSystem to Toil (resolves #796)
Adds a YARNBatchSystem that enables Toil to run jobs on a cluster that is
managed with the Apache Hadoop YARN resource manager. A job maps directly to a
YARN application. Specifically, we run a batch command line under YARN using
the YARN DistributedShell application. This way, we do not need to implement
a YARN ApplicationManager, etc.