Synchronize cwltool version with arvados-cwl-runner
Moves cwltool version pin to match that used in arvados-cwl-runner so
both can exist in the same python environment. This has been tested with
full bcbio generated CWL runs on arvados-cwl-runner and cwltoil.
Discussed with @tetron who recommended synchronizing until we move to
semantic versioning on cwltool. Follows up on #910.