Create test to replicate #966
Refactored asyncWrite to be a method in FileStore to be a method in FileStore instead of a function in Filestore.__init__ .
Wrote a test to replicate #966. This commit modifies jenkins.sh to show the testfailing across batch systems. A subsequent commit will add the code that will show the issue as fixed.

Fixed some minor typos too.