Improve logging message for filestore CWL runs
The output logging messages when running cwltoil pipelines were generic
references to jobstore directories, making it hard for users to
determine which stage of the pipeline was running. This fix
handles the case where we're running CWL-based runs with filesystem
storage, reading the underlying CWL and providing a useful logging
message from the baseCommand.

This is a temporary fix until a more general system can be put in place.