CWL: Pass local temporary directory to jobs
This fixes problems on shared filesystem batch scheduler systems which
do not automatically transfer TMPDIR from the parent. Failing to do this
results in jobs run on non-shared local temporary directories which are
not accessible to the parent fileStorage manager. This avoids the issue
by explicitly setting and passing a temporary directory in the defined
outdir. Required for distributed runs on SGE clusters so supplements #1105
with the necessary CWL specific fixes.