Implement support for deferred functions (resolves #987)
resolves #987
related to #922

Jobs can now register deferred functions -- functions that are run when the
job finishes, when the job crashes and is restarted or upon batch system
shutdown.
Every job now tests the state of other jobs on the node and can handle
cleanup of other jobs that have failed.
Added 4 tests for the deferred functions (it works with methods, class methods,
and lambdas, it works on restart, jobs can handle other dead jobs, worker
cleanup handles dead jobs)