Add disk-usage logging to filestores (resolves #1188)
resolves #1188

Added a simple check at the end of each job of how much disk space is used by the
job at the end of its lifespan. Log the results to INFO and then add a message to
WARNING if the job used more space than it requested.
Added a function to recursively get the size of a directory (including subdirectories).
Added a test for the function under a new test file `src/miscTests.py`