Add convenience tags to appliance image (resolves #1244)
Decided not to implement tagging with `latest` as it’s very difficult to determine which release should be deemed `latest`. If there are two release branches, one for the latest stable release and one for an older legacy release, only images for the former should be tagged `latest`. This is impossible to decide without listing all release branches which is something I decided not to do for the moment.