Resolves #1222, resolves #1289, incorporates #1292.
A job with multiple predecessors and one or more failing predecessors
could cause a potential deadlock scenario. 

Furthermore, restarting a workflow containing a job with multiple
predecessors not all of which are complete could lead to the job being
run before all its predecessors were succesfully completed. 

Both these bugs are squashed by this PR, which adds the proper
tracking/restart of jobs with multiple predecessors under scenarios in
which one or more predecessors of such a job fails.