Disable pytest’s fancy tracebacks and assertion introspection
The tracebacks are too verbose and often fail to inline the relevant sections
since it’s not a given that the top and bottom of a traceback are the most
relevant. It's a heuristic that fails too often to offset the impact of the
verbosity.

I've seen the assertion introspection fail in certain cases, e.g. printing a
repr() of a method rather than the repr() of that method's return value.