CWL: symlink jobStore files instead of copying
getFile would copy all input files from the jobStore previous due to a
refactoring bug: the `copy` function argument was removed but still
referenced; it picked up the `copy` module and was always true.

This moves to using symlinks to avoid the extra file copy. In doing so
it exposed some new issues, also fixed in this PR:

- Needed to correctly set basename and location in the File objects to
  avoid resolving the original Toil name instead of the actual file name.

- We don't support writable InitialWorkDirRequirement in CWL, but still
  support the more common readable case.

Fixes #1627