Add test for Directory
Added a test for a simple cwl step that that uses a Directory as input. The
test is failing as expected, but under "Captured stderr call" we can see the
logging of cwl and in particular the ERROR "CWL Directory inputs not yet
supported in Toil". This shows that cwl is correctly called and fails because
the feature under development has not been implemented yet