Fix docker tests failing on mac. resolves #1778
Docker on mac requires whitelisting directories before mounting them into docker images. Temp files are created in `/var/folders/...`. `/var` is a symlink to `/var/private`. `/private` is in the whitelist, but since the canonical name is not being used docker doesn't recognize it.

The workaround of adding `/var/folders` fails because the GUI will only select the canonical name (`/private/var/folders`) and because `/private` is already on the list the change is rejected.

resolves #1778