Stop log level option defaults from clobbering each other
The default default of None for some of the options with a dest of
"logLevel" was overriding the explicitly set default. Adding
an explicit default to all of them solves the problem.

Also restores functionality of the --logOff option which didn't
work.

Fixes #1316 by having a non-None logLevel for --realTimeLogging to
use.

(resolves #1316)