CWL: run internal jobs locally, not on cluster
The cwltoil integration includes non-processing based internal
jobs that handle dispatching and managing workflows. When running on a
cluster these jobs were being submitted to the scheduler. The larger
number of short submissions make busy schedulers unhappy and slow down
the runs.

This introduces a local batch queue along with the cluster submission
queue. Internal CWL jobs go to this, getting run locally instead of on
the cluster. It's a general fix that should apply across all of the
supported batch systems (SGE, Slurm, Torque, PBSPro).