Put job name into the _toil_worker arguments
The job name isn't used for anything important, but it's very useful
if you're trying to do a quick ps to see what's running. Previously
only the jobStore and jobStoreID was passed. This should make it much
easier to see at a glance what is running on any particular node.

Closes #1884.