Batch: cache instead of block for statePollingWait
Improves support for configuring interval to query schedulers. This
builds off #1769 which blocked to avoid extra queries. Unfortunately
this also blocks submissions of jobs, making it hard to scale up
analyses. This change uses caching instead of blocking, so Toil can
continue to submit and monitor jobs and will refresh queries to the
scheduler only within the `statePollingWait` period. The behavior when
`statePollingWait` is unset does not change.