Fix sort tests to reflect new restart-completed-job-store behavior
Previously it assumed that a completed jobStore (when intentionally
kept around with clean="never") would raise a JobException on restart
because of the missing root job. That isn't true anymore because we
want to allow exports to be re-attempted.

This also fixes a few weird bugs in the sort test lingering from the
recent sort test changes. The output file was hardcoded, and the
output was never actually checked to be sorted correctly.