Fix launch-cluster requiring your terminal window to be wide
Toil uses ssh -t to issue commands, which forces the allocation of a
TTY. When checking that dockerd is running, it issues a ps aux and
searches for dockerd in the output. But when run inside a TTY, ps
truncates the output to the width of the terminal. If you used the
default terminal width, 80, and were particularly unlucky with what
date/PID/RSS you had, you could get an output for the dockerd process
that looks like this:

root      1002  8.3  1.8 816052 73688 ?        Ssl  18:37   0:30 /run/torcx/bin

which actually doesn't contain the string "dockerd" at all.

This patch just forces ps to never truncate.