Make ignoredNodes functionality actually work
The ignoredNodes functionality used to work on the basis of private
IPs, but the Mesos batch system checks against the ignoredNodes set on
the basis of hostnames. Except in very rare cases (when the
reverse-lookup that Mesos performs by default fails), those were
completely different values.

This patch forces Mesos workers to use the private IP when reporting
their "hostname".