Retry on instance ID NotFound errors when tagging
EC2 seems to have eventual consistency on some data like instance
IDs. Sometimes we are unlucky, and after we launch an instance, we
immediately attempt to tag it and our request gets routed to a
different server, one which hasn't had the instance ID propagated to
it yet, leaving us with a 400 response.

This is a common error which a lot of people have run into. The root
cause is that boto2 doesn't actually support creating tags on instance
creation, though EC2 does. Spark ran into the same issue several years
ago (https://issues.apache.org/jira/browse/SPARK-4983) and decided to
solve it by sleeping for 5 to 30 seconds after instance creation, but
that won't be acceptable for our purposes. This just retries on those
specific errors when tagging.

The proper solution to this problem would be to switch to boto3 for
launching instances, since it supports adding tags at creation time
without extra requests. Since that's a bit of a bigger change, we'll
leave that for later and try to stem the immediate problem for now.