Put all thread logic in ScalerThread, all scaler logic in ClusterScaler
Previously these were reversed. The original reason was that we would
actually have two scaler threads for every "scaler", one preemptable
and one non-preemptable. But after we pruned this down to a single
thread, the hierarchy where ClusterScaler created the scaler thread,
but passed itself to the child thread, was pretty weird.