Add option not to generate a master public key (resolves #2154)
Add an option to Toil to not create a master public key for the
provisioner when using autoscaling and running the pipeline from
Dockstore (which calls cwl-runner) where the pipeline is running
as the leader node, because cwl-runner locks the file system on
the leader and creation of an ssh key is not possible. If the
provisioner attempts to create the ssh key an exception will be thrown.
Resolves: #2154