Test Kubernetes when appliance ought to be in Quay
We already had a `needs_appliance` which checks for the appliance being cached in the local Docker, which isn't what we want here.

TODO: How do we avoid running this test for local testing when you have Kubernetes but you haven't pushed to Quay? Should we really be polling Quay?