Fix cpu_count() to actually read quotas (#2933)
* Log more information about CPU count finding

* Make cpu_count actually read files, and cache

* Make sure all tests have AWS and Kubernetes access

* Log more information for failed pods caught by watch

* Don't run Kubernetes tests on Python 2 where there is no appliance

* Don't make a second needs_appliance

* Test Kubernetes when appliance ought to be in Quay

We already had a `needs_appliance` which checks for the appliance being cached in the local Docker, which isn't what we want here.

TODO: How do we avoid running this test for local testing when you have Kubernetes but you haven't pushed to Quay? Should we really be polling Quay?

* Include explainations when failing script tests

Now the assertion failures should include the observed and expected output.