Raise default min CPU size of single machine jobs
This should stop us spawning quite so many threads on single machine, and help with #2928. The cost is that jobs that ask for less than a core will get an entire core to themselves. A real solution without all these placeholder threads would address that, but this change is easy and should be a good stopgap.