Deduplicate updated jobs
As is, multiple JobGraphs for the same job ID can be updated in the same tick.
I don't think multiple job graphs for the same job ID are supposed to exist at
all.

I'm now deduplicating everything by job ID. THe first update with the first
status will win.

Really we should change the whole system to pass around just the IDs of updated
things, and store a single source of truth for the job graph's state, so it is
impossible to have conflicting updates in flight.