Mark CWL tests on Kupernetes as expected to fail
Around 20-40 of the CWL tests are intermittently failing on Kubernetes, and
it's not worrth re-running the tests over and over to get all their numbers.

Common failure modes include missing ordering constraints, and missing local
file paths.

This lets all the tests run (so we can see which are still broken), but expects
the runs to fail. When we actually fix a test, we can add it to another test
cakll that isn't xfail and runs it alone with selected_tests. Or, we can try
and fix them all in a row and hope there's no backsliding, and then remove
xfail.