Pickle toilState to pass to workers for jobchaining work.
The parent thread now pickle's toilState, passing
it to the worker thread, where it is unpickled again.

A test for completed predecessors for jobs with multiple
predecessors has been added to the worker, so that some
basic job chaining can be performed for jobs whose remaining
predecessor completes within that task.

This change does result in the parent thread not correctly
noting jobs which had multiple predecessors have now been
completed correctly. So some of the checks at the end of
the internal job routine have had to be switched off for the
moment.