Fix bug and add doc strings
The Job IDs as provided in `batch_job_id_list` do not contain integers, but
strings, containing "<job-id>[.<job-task>]". This bug has been fixed.
Doc strings were added to the new and modified functions.

TODO: See if we should use keep "<job-id>" part as string instead, to prevent
conversion to integer, which might fail if, for whatever reason, "<job-id>"
cannot be converted to an integer.