Consolidated CI Fixing (#3887)
* Adjust testExtremeCacheSetup failure constants

This should fix #3886 by making testExtremeCacheSetup more likely to succeed faster. It might be a less good test of caching robustness, but it's no help for that if it's timing out when nothing in caching has changed.

* Force ruamel packages past a pickle-breaking change

Older Toil installs can't talk to new containers unless they have at least
these versions. The container tries to unpickle the old module's data
structures and produces:

```
Traceback (most recent call last):
  File "/usr/local/lib/python3.6/dist-packages/toil/worker.py", line 376, in workerScript
    job = Job.loadJob(jobStore, jobDesc)
  File "/usr/local/lib/python3.6/dist-packages/toil/job.py", line 2251, in loadJob
    job = cls._unpickle(userModule, fileHandle, requireInstanceOf=Job)
  File "/usr/local/lib/python3.6/dist-packages/toil/job.py", line 1876, in _unpickle
    runnable = unpickler.load()
AttributeError: 'Comment' object has no attribute '_end'
```

* Hint Pip what ruamel.yaml will work with our cwltool

* Fix #3908 by adding some retry to the Singularity wrapper

* Fix #3894 by waiting around for longer trying to save multipart uploads

* Use junit output to actually find output for failing tests

* Quote odd YAML strings

* Add timeout, name services, and detect vanished

This should help with #3910 by not letting the test deadlock, because of the
timeout, and because we now refuse to do a blocking wait on services not
actually in the job store.

* Get back vaguely good names for function wrapping jobs

* Don't let services finish early

This makes sure to wait for the leader to see that child jobs are done, and ask
services to stop, before letting the services stop successfully.

This should fix https://github.com/DataBiosphere/toil/issues/3484

* Make random job graph test fast enough now that services have time to be killed

* Adjust deadlock detector to not ignore jobs that run and succeed/fail between checks

* Lower core requirements to prevent actual deadlocks

* Also lower memory and disk

* Use a CWL test runner that will keep output for timed-out tests

* Make extra-sure statuses are numbers

This may fix https://github.com/DataBiosphere/toil/issues/3913 by excluding the possibility of comparing a string HTTP status to a number.

I'm not sure if that ever could happen, but I don't have a better theory as to why the request in that issue wouldn't be retried.

* Don't overload the leader with CWL tests

* Stop using up Dacker Hub pulls when testing our image checker

* Temporarily drop all tests except the CWL Kubernetes tests

* Revert "Temporarily drop all tests except the CWL Kubernetes tests"

This reverts commit 91ee104cc654d8ae5cc60aa2f7455792c9b2ff29.

* Remove xfail from main branch.

* Pin package; prevent potential pickle problems

* Improve spelling

Co-authored-by: Lon Blauvelt <lblauvel@ucsc.edu>