workDir and jobStore now default to (shared) tmp-outdir-prefix (#5154)
* Let workDir and jobStore default to tmp-outdir-prefix

This commit makes the following changes to the behaviour of the given command-line options:
* `tmp-outdir-prefix` defaults to `tmpdir-prefix`, unless given on the command-line
* `workDir` defaults to `tmp-outdir-prefix`, unless given on the command-line
* `jobStore` defaults to `tmp-outdir-prefix`, unless given on the command-line
* `coordinationDir` defaults to the default tmpdir-prefix, ignoring `tmpdir-prefix` when given on the command-line (rationale: this is a book-keeping location, that must be on a local 100% posix-compliant file system, because it uses file locks).

* Update documentation

Updated the CLI documention for the `--workDir` option.

* Jobstore needs its own directory

The jobstore cannot be put inside the working directory, because it may need to be retained (e.g. when --stats is set).
It now gets its own (temporary) directory, if not specified with the --jobstore option.

* Do not create working directory on head node

We do not need to create a working directory on the head node. We only need to create our jobstore here.

* No need to set options.tmp_outdir_prefix

There's no need to set `options.tmp_outdir_prefix` here. It is not done in the current `master` branch either.

* Improve documentation of workDir option

Improved the documentation of the `--workDir` option, by adding an explanation that the `--tmp-outdir-prefix` will be used for CWL workflows.

* Update comment

---------

Co-authored-by: stxue1 <122345910+stxue1@users.noreply.github.com>
Co-authored-by: stxue1 <stxue@ucsc.edu>