Handle coercing file: URIs to files (#5266)
* Try to make sure setup.py can import the version template during codegen

Don't forget to import sys

* Handle coercing file: URIs to files

This should fix #5265.

Broad likes to write workflows that coerce Files to Strings and back
again. Apparently when we go File to String we always include a file:
URI prefix, and the String-to-FIle coercion was thinking that was a
nonexistent bare path, and also the caching code was not dealing with a
file: URI that didn't have a known cache path.

This fixes both of these issues at workflow scope.

At task scope we really shouldn't be going from String to File for
anything except what's on the task's filesystem; there might still be
dragons there because I'm not sure this is really allowed there.

* Adjust to new test setup

* Fix rsync detector for Mac rsync

* Make _virtualize_filename responsible for existence checking so it can use its symlink logic

* Make Docker nvidia check stop logging to console

* Move import to the top

* Fix spelling