Change from excluding unit tests by number to excluding by ID
This was accomplished with the "help" of Anthropic Claude, who did this
OK but then could not comprehend reading an entire SPEC.md diff to sniff
for important test changes after 10 minutes of badgering. And also it
took like 5 minutes of basgerign to get it to the grep | nl command I
was reaching for when I tried the robot instead.