Remove robot mistakes
Clean up after Anthropic Claude mis-registered the test numbers when
replacing the 1.1 tests withn names. We needed to use the test numbers
from the *old* build of the conformance test importer, and also the old
1.1 code.

Also use a version of the conformance testing repo that has its patch
data fixed up for the new test IDs that include _task, etc.